requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
ping3==4.0.8
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
urllib3==2.3.0
